+++ src/relay/blockchain/proxy.py
@@ -205,8 +205,11 @@ class Proxy(object):
 
     def _register_raw_event_log(self, raw_event_log) -> None:
         """Registers a new log to be decoded and sent to the correct event listener"""
-        log = self._get_processed_log_from_raw_log(raw_event_log)
-        self._event2log_queue[log["event"]].put(log)
+        try:
+            log = self._get_processed_log_from_raw_log(raw_event_log)
+            self._event2log_queue[log["event"]].put(log)
+        except AbiNotFoundException:
+            logger.warning(f"Got event with no matching abi: {raw_event_log}")
 
     def _get_processed_log_from_raw_log(self, raw_event_log) -> EventData:
         event_abi = self._get_abi_for_log(raw_event_log)