+++ tests/chain_integration/conftest.py
@@ -61,16 +61,10 @@ def taker(accounts):
 
 class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
     def setup_trustlines(self, trustlines):
-        for (A, B, clAB, clBA) in trustlines:
-            txid = self._proxy.functions.setAccount(
-                A, B, clAB, clBA, 0, 0, False, 0, 0
-            ).transact()
-            self._web3.eth.waitForTransactionReceipt(txid)
-
-    def setup_trustlines_with_interests_with_updates(self, trustlines_with_interests):
-        # uses `update_trustline` instead of `setAccount` so that events are emitted when setting up trustlines
-        for (A, B, clAB, clBA, intAB, intBA) in trustlines_with_interests:
-            self.update_trustline_with_accept(A, B, clAB, clBA, intAB, intBA)
+        for (A, B, clAB, clBA, intAB, intBA, is_frozen, balanceAB) in trustlines:
+            self.update_trustline_with_accept(
+                A, B, clAB, clBA, intAB, intBA, is_frozen, balanceAB
+            )
 
     def update_trustline(
         self,
@@ -81,10 +75,20 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
         interest_rate_given=None,
         interest_rate_received=None,
         is_frozen=False,
+        transfer=0,
     ):
         if interest_rate_given is None or interest_rate_received is None:
             txid = self._proxy.functions.updateCreditlimits(
                 to, creditline_given, creditline_received
+            ).transact({"from": from_})
+        elif transfer == 0:
+            txid = self._proxy.functions.updateTrustline(
+                to,
+                creditline_given,
+                creditline_received,
+                interest_rate_given,
+                interest_rate_received,
+                is_frozen,
             ).transact({"from": from_})
         else:
             txid = self._proxy.functions.updateTrustline(
@@ -94,8 +98,10 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
                 interest_rate_given,
                 interest_rate_received,
                 is_frozen,
+                transfer,
             ).transact({"from": from_})
         self._web3.eth.waitForTransactionReceipt(txid)
+        return txid
 
     def cancel_trustline_update(self, from_, to):
         txid = self._proxy.functions.cancelTrustlineUpdate(to).transact({"from": from_})
@@ -110,6 +116,7 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
         interest_rate_given=None,
         interest_rate_received=None,
         is_frozen=False,
+        transfer=0,
     ):
         self.update_trustline(
             from_,
@@ -119,8 +126,9 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
             interest_rate_given,
             interest_rate_received,
             is_frozen,
+            transfer,
         )
-        self.update_trustline(
+        return self.update_trustline(
             to,
             from_,
             creditline_received,
@@ -128,6 +136,7 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
             interest_rate_received,
             interest_rate_given,
             is_frozen,
+            -transfer,
         )
 
     def update_trustline_and_cancel(
@@ -171,6 +180,18 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
             is_frozen,
         )
         self.cancel_trustline_update(to, from_)
+
+    def close_trustline(self, from_, to):
+        self._proxy.functions.closeTrustline(to).transact({"from": from_})
+
+    def settle_and_close_trustline(self, from_, to):
+        balance = self.get_balance(from_, to)
+        if balance > 0:
+            self.transfer(from_, balance, max_fee=0, path=[from_, to])
+        elif balance < 0:
+            self.transfer(to, -balance, max_fee=0, path=[to, from_])
+
+        self.close_trustline(from_, to)
 
     def transfer(self, from_, value, max_fee, path, extra_data=b""):
         tx_id = self._proxy.functions.transfer(
@@ -231,28 +252,28 @@ class CurrencyNetworkProxy(currency_network_proxy.CurrencyNetworkProxy):
 @pytest.fixture(scope="session")
 def trustlines(accounts):
     return [
-        (accounts[0], accounts[1], 100, 150),
-        (accounts[1], accounts[2], 200, 250),
-        (accounts[2], accounts[3], 300, 350),
-        (accounts[3], accounts[4], 400, 450),
-        (accounts[4], accounts[5], 400, 450),
-        (accounts[5], accounts[6], 400, 450),
-        (accounts[0], accounts[4], 500, 550),
-    ]  # (A, B, clAB, clBA)
+        (accounts[0], accounts[1], 100, 150, 0, 0, False, 0),
+        (accounts[1], accounts[2], 200, 250, 0, 0, False, 0),
+        (accounts[2], accounts[3], 300, 350, 0, 0, False, 0),
+        (accounts[3], accounts[4], 400, 450, 0, 0, False, 0),
+        (accounts[4], accounts[5], 400, 450, 0, 0, False, 0),
+        (accounts[5], accounts[6], 400, 450, 0, 0, False, 0),
+        (accounts[0], accounts[4], 500, 550, 0, 0, False, 0),
+    ]  # (A, B, clAB, clBA, intAB, intBA, frozen, transferAB)
 
 
 @pytest.fixture(scope="session")
 def trustlines_with_interests(accounts):
     return [
-        (accounts[0], accounts[1], 12345, 12345, 2000, 1000),
-        (accounts[1], accounts[2], 12345, 12345, 2000, 1000),
-        (accounts[2], accounts[3], 12345, 12345, 2000, 1000),
-        (accounts[3], accounts[4], 12345, 12345, 2000, 1000),
-        (accounts[0], accounts[4], 12345, 12345, 2000, 1000),
-    ]  # (A, B, clAB, clBA)
+        (accounts[0], accounts[1], 12345, 12345, 2000, 1000, False, 0),
+        (accounts[1], accounts[2], 12345, 12345, 2000, 1000, False, 0),
+        (accounts[2], accounts[3], 12345, 12345, 2000, 1000, False, 0),
+        (accounts[3], accounts[4], 12345, 12345, 2000, 1000, False, 0),
+        (accounts[0], accounts[4], 12345, 12345, 2000, 1000, False, 0),
+    ]  # (A, B, clAB, clBA, intAB, intBA, frozen, transferAB)
 
 
-def deploy_test_network(web3):
+def deploy_currency_network_v2(web3):
     return deploy_network(
         web3,
         network_settings=NetworkSettings(
@@ -262,7 +283,7 @@ def deploy_test_network(web3):
             custom_interests=True,
             expiration_time=EXPIRATION_TIME,
         ),
-        currency_network_contract_name="TestCurrencyNetwork",
+        currency_network_contract_name="CurrencyNetworkV2",
     )
 
 
@@ -276,8 +297,8 @@ def contracts():
 
 
 @pytest.fixture(scope="session")
-def currency_network_abi(contracts):
-    return contracts["TestCurrencyNetwork"]["abi"]
+def currency_network_v2_abi(contracts):
+    return contracts["CurrencyNetworkV2"]["abi"]
 
 
 @pytest.fixture(scope="session")
@@ -292,22 +313,22 @@ def token_abi(contracts):
 
 @pytest.fixture(scope="session")
 def testnetwork1_address(web3):
-    return deploy_test_network(web3).address
+    return deploy_currency_network_v2(web3).address
 
 
 @pytest.fixture(scope="session")
 def testnetwork2_address(web3):
-    return deploy_test_network(web3).address
+    return deploy_currency_network_v2(web3).address
 
 
 @pytest.fixture(scope="session")
 def testnetwork3_address(web3, chain):
-    return deploy_test_network(web3).address
+    return deploy_currency_network_v2(web3).address
 
 
 @pytest.fixture(scope="session")
 def testnetwork4_address(web3):
-    return deploy_test_network(web3).address
+    return deploy_currency_network_v2(web3).address
 
 
 @pytest.fixture()
@@ -347,19 +368,19 @@ def network_addresses_with_exchange(testnetworks):
 
 
 @pytest.fixture(scope="session")
-def currency_network(web3, currency_network_abi, testnetwork1_address):
+def currency_network(web3, currency_network_v2_abi, testnetwork1_address):
     currency_network = CurrencyNetworkProxy(
-        web3, currency_network_abi, testnetwork1_address
+        web3, currency_network_v2_abi, testnetwork1_address
     )
     return currency_network
 
 
 @pytest.fixture()
 def currency_network_with_trustlines(
-    web3, currency_network_abi, testnetwork2_address, trustlines
+    web3, currency_network_v2_abi, testnetwork2_address, trustlines
 ):
     currency_network = CurrencyNetworkProxy(
-        web3, currency_network_abi, testnetwork2_address
+        web3, currency_network_v2_abi, testnetwork2_address
     )
     currency_network.setup_trustlines(trustlines)
 
@@ -368,38 +389,34 @@ def currency_network_with_trustlines(
 
 @pytest.fixture()
 def currency_network_with_trustlines_and_interests(
-    web3, currency_network_abi, testnetwork2_address, trustlines_with_interests
+    web3, currency_network_v2_abi, testnetwork2_address, trustlines_with_interests
 ):
     currency_network = CurrencyNetworkProxy(
-        web3, currency_network_abi, testnetwork2_address
+        web3, currency_network_v2_abi, testnetwork2_address
     )
-    currency_network.setup_trustlines_with_interests_with_updates(
-        trustlines_with_interests
-    )
+    currency_network.setup_trustlines(trustlines_with_interests)
 
     return currency_network
 
 
 @pytest.fixture(scope="session")
 def currency_network_with_trustlines_and_interests_session(
-    web3, currency_network_abi, testnetwork3_address, trustlines_with_interests
+    web3, currency_network_v2_abi, testnetwork3_address, trustlines_with_interests
 ):
     currency_network = CurrencyNetworkProxy(
-        web3, currency_network_abi, testnetwork3_address
+        web3, currency_network_v2_abi, testnetwork3_address
     )
-    currency_network.setup_trustlines_with_interests_with_updates(
-        trustlines_with_interests
-    )
+    currency_network.setup_trustlines(trustlines_with_interests)
 
     return currency_network
 
 
 @pytest.fixture(scope="session")
 def currency_network_with_trustlines_session(
-    web3, currency_network_abi, testnetwork4_address, trustlines
+    web3, currency_network_v2_abi, testnetwork4_address, trustlines
 ):
     currency_network = CurrencyNetworkProxy(
-        web3, currency_network_abi, testnetwork4_address
+        web3, currency_network_v2_abi, testnetwork4_address
     )
     currency_network.setup_trustlines(trustlines)