+++ tests/chain_integration/database_integration/test_events_information.py
@@ -10,7 +10,10 @@ from relay.ethindex_db.events_informations import (
     IdentifiedNotPartOfTransferException,
 )
 from relay.network_graph.graph import CurrencyNetworkGraph
+from relay.network_graph.interests import calculate_interests
 from relay.network_graph.payment_path import FeePayer
+
+ONE_YEAR_IN_SECONDS = 365 * 3600 * 24
 
 
 def accrue_interests(currency_network, web3, chain, path, years):
@@ -20,7 +23,7 @@ def accrue_interests(currency_network, web3, chain, path, years):
 
     for year in years:
         timestamp = web3.eth.getBlock("latest").timestamp
-        timestamp += (3600 * 24 * 365) * year + 1
+        timestamp += ONE_YEAR_IN_SECONDS * year + 1
 
         chain.time_travel(timestamp)
         chain.mine_block()
@@ -28,7 +31,7 @@ def accrue_interests(currency_network, web3, chain, path, years):
 
 
 @pytest.mark.parametrize(
-    "years, interests", [([0, 1], [0, 2]), ([1, 4, 2, 3], [1, 9, 8, 19])]
+    "years, interests", [([0, 1], [2]), ([1, 4, 2, 3], [1, 9, 8, 19])]
 )
 def test_get_interests_received_for_trustline_positive_balance(
     ethindex_db_for_currency_network_with_trustlines_and_interests,
@@ -61,7 +64,7 @@ def test_get_interests_received_for_trustline_positive_balance(
 
 
 @pytest.mark.parametrize(
-    "years, interests", [([0, 1], [0, 4]), ([1, 4, 2, 3], [2, 24, 26, 74])]
+    "years, interests", [([0, 1], [4]), ([1, 4, 2, 3], [2, 24, 26, 74])]
 )
 def test_get_interests_received_for_trustline_negative_balance(
     ethindex_db_for_currency_network_with_trustlines_and_interests,
@@ -94,7 +97,7 @@ def test_get_interests_received_for_trustline_negative_balance(
 
 
 @pytest.mark.parametrize(
-    "years, interests", [([0, 1], [0, -2]), ([1, 4, 2, 3], [-1, -9, -8, -19])]
+    "years, interests", [([0, 1], [-2]), ([1, 4, 2, 3], [-1, -9, -8, -19])]
 )
 def test_get_interests_paid_for_trustline_positive_balance(
     ethindex_db_for_currency_network_with_trustlines_and_interests,
@@ -127,7 +130,7 @@ def test_get_interests_paid_for_trustline_positive_balance(
 
 
 @pytest.mark.parametrize(
-    "years, interests", [([0, 1], [0, -4]), ([1, 4, 2, 3], [-2, -24, -26, -74])]
+    "years, interests", [([0, 1], [-4]), ([1, 4, 2, 3], [-2, -24, -26, -74])]
 )
 def test_get_interests_paid_for_trustline_negative_balance(
     ethindex_db_for_currency_network_with_trustlines_and_interests,
@@ -157,6 +160,184 @@ def test_get_interests_paid_for_trustline_negative_balance(
     assert list_of_interests == interests
     for accrued_interest in accrued_interests:
         assert accrued_interest.interest_rate == 2000
+
+
+@pytest.mark.parametrize(
+    "years, interests", [([0, 1], [2]), ([1, 4, 2, 3], [1, 9, 8, 19])]
+)
+def test_get_interests_received_open_trustline_with_transfer(
+    ethindex_db_for_currency_network_with_trustlines_and_interests,
+    currency_network_with_trustlines_and_interests_session,
+    web3,
+    chain,
+    accounts,
+    years,
+    interests,
+    wait_for_ethindex_to_sync,
+):
+    """Test that opening a trustline with balance will still allow us to get interests"""
+    currency_network = currency_network_with_trustlines_and_interests_session
+
+    currency_network.close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[1], accounts[2], 12345, 12345, 0, 0, False, 123
+    )
+    currency_network.transfer(accounts[2], 123, 0, [accounts[2], accounts[1]])
+
+    currency_network.update_trustline_with_accept(
+        accounts[1], accounts[2], 12345, 12345, 2000, 1000, False, 0
+    )
+
+    path = [accounts[0], accounts[1], accounts[2], accounts[3]]
+    accrue_interests(currency_network, web3, chain, path, years)
+    wait_for_ethindex_to_sync()
+
+    accrued_interests = EventsInformationFetcher(
+        ethindex_db_for_currency_network_with_trustlines_and_interests
+    ).get_list_of_paid_interests_for_trustline(
+        currency_network.address, accounts[2], accounts[1]
+    )
+    list_of_interests = [
+        accrued_interest.value for accrued_interest in accrued_interests
+    ]
+    assert list_of_interests == interests
+    for accrued_interest in accrued_interests:
+        assert accrued_interest.interest_rate == 1000
+
+
+@pytest.mark.parametrize(
+    "years, interests", [([0, 1], [4]), ([1, 4, 2, 3], [2, 24, 26, 74])]
+)
+def test_get_interests_paid_open_trustline_with_transfer(
+    ethindex_db_for_currency_network_with_trustlines_and_interests,
+    currency_network_with_trustlines_and_interests_session,
+    web3,
+    chain,
+    accounts,
+    years,
+    interests,
+    wait_for_ethindex_to_sync,
+):
+    """Test that opening a trustline with balance will still allow us to get paid interests"""
+    currency_network = currency_network_with_trustlines_and_interests_session
+
+    currency_network.close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[1], accounts[2], 12345, 12345, 0, 0, False, 123
+    )
+    currency_network.transfer(accounts[2], 123, 0, [accounts[2], accounts[1]])
+
+    currency_network.update_trustline_with_accept(
+        accounts[1], accounts[2], 12345, 12345, 2000, 1000, False, 0
+    )
+
+    path = [accounts[3], accounts[2], accounts[1], accounts[0]]
+    accrue_interests(currency_network, web3, chain, path, years)
+    wait_for_ethindex_to_sync()
+
+    accrued_interests = EventsInformationFetcher(
+        ethindex_db_for_currency_network_with_trustlines_and_interests
+    ).get_list_of_paid_interests_for_trustline(
+        currency_network.address, accounts[1], accounts[2]
+    )
+    list_of_interests = [
+        accrued_interest.value for accrued_interest in accrued_interests
+    ]
+    assert list_of_interests == interests
+    for accrued_interest in accrued_interests:
+        assert accrued_interest.interest_rate == 2000
+
+
+def test_get_interests_received_during_opening_of_trustline(
+    ethindex_db_for_currency_network_with_trustlines_and_interests,
+    currency_network_with_trustlines_and_interests_session,
+    web3,
+    chain,
+    accounts,
+    wait_for_ethindex_to_sync,
+):
+    """Test that opening a trustline with balance and interests will allow us to get resulting accrued interests"""
+    currency_network = currency_network_with_trustlines_and_interests_session
+    initial_transfer = 100
+    interest_rate = 1000
+
+    currency_network.close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[1],
+        accounts[2],
+        12345,
+        12345,
+        2000,
+        interest_rate,
+        False,
+        initial_transfer,
+    )
+
+    path = [accounts[0], accounts[1], accounts[2], accounts[3]]
+    timestamp = web3.eth.getBlock("latest").timestamp
+    timestamp += ONE_YEAR_IN_SECONDS + 1
+    chain.time_travel(timestamp)
+    currency_network.transfer(path[0], 9, 1000, path, b"")
+
+    wait_for_ethindex_to_sync()
+
+    accrued_interests = EventsInformationFetcher(
+        ethindex_db_for_currency_network_with_trustlines_and_interests
+    ).get_list_of_paid_interests_for_trustline(
+        currency_network.address, accounts[2], accounts[1]
+    )
+    assert len(accrued_interests) == 1
+    assert accrued_interests[0].value == calculate_interests(
+        initial_transfer, interest_rate, ONE_YEAR_IN_SECONDS
+    )
+    assert accrued_interests[0].interest_rate == interest_rate
+    assert accrued_interests[0].timestamp == web3.eth.getBlock("latest").timestamp
+
+
+def test_get_interests_paid_during_opening_of_trustline(
+    ethindex_db_for_currency_network_with_trustlines_and_interests,
+    currency_network_with_trustlines_and_interests_session,
+    web3,
+    chain,
+    accounts,
+    wait_for_ethindex_to_sync,
+):
+    """Test that opening a trustline with balance and interests will allow us to get resulting paid interests"""
+    currency_network = currency_network_with_trustlines_and_interests_session
+    initial_transfer = -100
+    interest_rate = 2000
+
+    currency_network.close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[1],
+        accounts[2],
+        12345,
+        12345,
+        interest_rate,
+        1000,
+        False,
+        initial_transfer,
+    )
+
+    path = [accounts[0], accounts[1], accounts[2], accounts[3]]
+    timestamp = web3.eth.getBlock("latest").timestamp
+    timestamp += ONE_YEAR_IN_SECONDS + 1
+    chain.time_travel(timestamp)
+    currency_network.transfer(path[0], 9, 1000, path, b"")
+
+    wait_for_ethindex_to_sync()
+
+    accrued_interests = EventsInformationFetcher(
+        ethindex_db_for_currency_network_with_trustlines_and_interests
+    ).get_list_of_paid_interests_for_trustline(
+        currency_network.address, accounts[2], accounts[1]
+    )
+    assert len(accrued_interests) == 1
+    assert accrued_interests[0].value == calculate_interests(
+        initial_transfer, interest_rate, ONE_YEAR_IN_SECONDS
+    )
+    assert accrued_interests[0].interest_rate == interest_rate
+    assert accrued_interests[0].timestamp == web3.eth.getBlock("latest").timestamp
 
 
 class LookupMethod(Enum):
@@ -273,6 +454,66 @@ def test_get_transfer_information_values(
     assert transfer_information.value == value
     assert transfer_information.total_fees == number_of_mediators
     assert transfer_information.fee_payer == fee_payer
+    assert transfer_information.currency_network == network.address
+
+
+@pytest.mark.parametrize("lookup_method", list(LookupMethod))
+@pytest.mark.parametrize("transfer_value", [100, -100])
+def test_get_transfer_information_trustline_update(
+    ethindex_db_for_currency_network_with_trustlines,
+    currency_network_with_trustlines_session,
+    web3,
+    accounts,
+    lookup_method,
+    transfer_value,
+    wait_for_ethindex_to_sync,
+):
+    """
+    test that we can get the transfer information from a transfer applied when opening a trustline
+    """
+    network = currency_network_with_trustlines_session
+
+    initiator = accounts[1]
+    counterparty = accounts[2]
+
+    network.close_trustline(initiator, counterparty)
+    block_number = web3.eth.blockNumber
+    tx_hash = network.update_trustline_with_accept(
+        initiator, counterparty, 12345, 12345, 0, 0, False, transfer_value
+    ).hex()
+
+    wait_for_ethindex_to_sync()
+
+    if (
+        lookup_method == LookupMethod.TRANSFER_ID
+        or lookup_method == LookupMethod.BALANCE_UPDATE_ID
+    ):
+        if lookup_method == LookupMethod.TRANSFER_ID:
+            event = network._proxy.events.Transfer().getLogs(fromBlock=block_number)[0]
+        elif lookup_method == LookupMethod.BALANCE_UPDATE_ID:
+            event = network._proxy.events.BalanceUpdate().getLogs(
+                fromBlock=block_number
+            )[0]
+        else:
+            assert False, "Unexpected lookup method"
+        transfer_information = EventsInformationFetcher(
+            ethindex_db_for_currency_network_with_trustlines
+        ).get_transfer_details_for_id(event["blockHash"].hex(), event["logIndex"])[0]
+    elif lookup_method == LookupMethod.TX_HASH:
+        transfer_information = EventsInformationFetcher(
+            ethindex_db_for_currency_network_with_trustlines
+        ).get_transfer_details_for_tx(tx_hash)[0]
+    else:
+        assert False, "Unknown lookup method"
+
+    if transfer_value > 0:
+        assert transfer_information.path == [initiator, counterparty]
+    else:
+        assert transfer_information.path == [counterparty, initiator]
+    assert transfer_information.fees_paid == []
+    assert transfer_information.value == abs(transfer_value)
+    assert transfer_information.total_fees == 0
+    assert transfer_information.extra_data == b""
     assert transfer_information.currency_network == network.address
 
 
@@ -413,6 +654,60 @@ def test_get_mediation_fees_with_pollution(
     tx_hash3 = custom_make_transfer()
     timestamp3 = web3.eth.getBlock("latest")["timestamp"]
 
+    # Pollution from opening trustline with a transfer
+    open_trustline_transfer_value = 123
+
+    currency_network.settle_and_close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[1],
+        accounts[2],
+        12345,
+        12345,
+        0,
+        0,
+        False,
+        open_trustline_transfer_value,
+    )
+
+    currency_network.settle_and_close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[1],
+        accounts[2],
+        12345,
+        12345,
+        0,
+        0,
+        False,
+        -open_trustline_transfer_value,
+    )
+
+    currency_network.settle_and_close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[2],
+        accounts[1],
+        12345,
+        12345,
+        0,
+        0,
+        False,
+        open_trustline_transfer_value,
+    )
+
+    currency_network.settle_and_close_trustline(accounts[1], accounts[2])
+    currency_network.update_trustline_with_accept(
+        accounts[2],
+        accounts[1],
+        12345,
+        12345,
+        0,
+        0,
+        False,
+        -open_trustline_transfer_value,
+    )
+
+    tx_hash4 = custom_make_transfer()
+    timestamp4 = web3.eth.getBlock("latest")["timestamp"]
+
     wait_for_ethindex_to_sync()
     graph = CurrencyNetworkGraph(
         capacity_imbalance_fee_divisor=currency_network.capacity_imbalance_fee_divisor,
@@ -424,7 +719,7 @@ def test_get_mediation_fees_with_pollution(
         ethindex_db_for_currency_network_with_trustlines_and_interests
     ).get_earned_mediation_fees(accounts[1], graph)
 
-    assert len(fees) == 4
+    assert len(fees) == 5
     custom_assert_fee = functools.partial(
         assert_fee, value=fee_value, from_=accounts[0], to=accounts[2]
     )
@@ -432,6 +727,7 @@ def test_get_mediation_fees_with_pollution(
     custom_assert_fee(fee=fees[1], tx_hash=tx_hash1, timestamp=timestamp1)
     custom_assert_fee(fee=fees[2], tx_hash=tx_hash2, timestamp=timestamp2)
     custom_assert_fee(fee=fees[3], tx_hash=tx_hash3, timestamp=timestamp3)
+    custom_assert_fee(fee=fees[4], tx_hash=tx_hash4, timestamp=timestamp4)
 
 
 def get_debts_of_single_currency_network(
@@ -661,6 +957,33 @@ def test_get_total_sum_transferred_with_noise(
     wait_for_ethindex_to_sync()
     sum = EventsInformationFetcher(
         ethindex_db_for_currency_network_with_trustlines
+    ).get_total_sum_transferred(sender, receiver)
+
+    assert sum == value
+
+
+def test_get_total_sum_transferred_trustline_update(
+    ethindex_db_for_currency_network,
+    currency_network: CurrencyNetworkProxy,
+    accounts,
+    wait_for_ethindex_to_sync,
+):
+    """Test getting the sum transferred for transfer occurring while opening a trustline"""
+    sender = accounts[0]
+    receiver = accounts[1]
+    value = 123
+
+    currency_network.update_trustline_with_accept(
+        sender, receiver, 12345, 12345, 0, 0, False, value,
+    )
+    currency_network.settle_and_close_trustline(sender, receiver)
+    currency_network.update_trustline_with_accept(
+        sender, receiver, 12345, 12345, 0, 0, False, -value,
+    )
+
+    wait_for_ethindex_to_sync()
+    sum = EventsInformationFetcher(
+        ethindex_db_for_currency_network
     ).get_total_sum_transferred(sender, receiver)
 
     assert sum == value